Fix for clip gpu thanks to rvencu

+++ clip_filter.py
@@ -52,7 +52,8 @@ class CLIP:
         return im_dataset["image_features"], im_dataset["similarity"]
 
     def prob(self, image_features, text_features):
-        image_features = torch.as_tensor(image_features).to(device)
+        text_features = text_features.float()
+        image_features = torch.as_tensor(image_features).to(device, dtype=torch.float32)
         image_features /= image_features.norm(dim=-1, keepdim=True)
         text_features /= text_features.norm(dim=-1, keepdim=True)
